ported to rails 8 (#32)
As part of the work to make the API handle requests asynchronously, we
looked for a job queue runner and couldn't find a modern one that could
be easily used with Sinatra, had few additional dependencies and had
good job persistence (ie would not drop or repeat jobs if a runner
crashed). Because the ecosystem (solid_queue, good_job, delayed_job) of
job runners we might choose from are all highly coupled with Rails, we
decided to migrate from Sinatra to Rails (and adopt ActiveRecord &
ActiveJob.

Alternatives considered:
- Sidekiq: You have to pay for Sidekiq Pro to get job persistence, and
introduces a Redis requirement
- Resque: You have to use a third party plugin to get retryable jobs,
has weird behavior (orphaned jobs) if the parent process crashes, and
introduces a Redis requirement.
- Custom Sequel backend for Delayed Job: Feels like too much complexity
to maintain ourselves (even though it seems to be working well in CMR)

---------

Co-authored-by: Jenny Heath <jheath@codeforamerica.org>
Co-authored-by: Kevin Diale <kdiale@codeforamerica.org>
Co-authored-by: Anisha Ramnani <aramnani@codeforamerica.org>
Co-authored-by: standard-ruby-action[bot] <standard-ruby-action[bot]@users.noreply.github.com>